+++ experimental/traffic-portal/src/app/api/profile.service.spec.ts
@@ -14,7 +14,7 @@
  */
 import { HttpClientTestingModule, HttpTestingController } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
-import { ProfileExport, ProfileType } from "trafficops-types";
+import { ProfileExport, ProfileType , ResponseProfile} from "trafficops-types";
 
 import { ProfileService } from "./profile.service";
 
@@ -50,6 +50,16 @@ describe("ProfileService", () => {
 			name: "TRAFFIC_ANALYTICS",
 			type: ProfileType.TS_PROFILE
 		}
+	};
+
+	const parameter = {
+		configFile: "cfg.txt",
+		id: 10,
+		lastUpdated: new Date(),
+		name: "TestParam",
+		profiles: null,
+		secure: false,
+		value: "TestVal"
 	};
 
 	beforeEach(() => {
@@ -144,6 +154,101 @@ describe("ProfileService", () => {
 		expect(req.request.body).toBe(importProfile);
 		req.flush({response: importProfile.profile});
 		await expectAsync(responseP).toBeResolvedTo(importProfile.profile);
+	});
+		
+	it("sends requests multiple Parameters", async () => {
+		const responseParams = service.getParameters();
+		const req = httpTestingController.expectOne(`/api/${service.apiVersion}/parameters`);
+		expect(req.request.method).toBe("GET");
+		expect(req.request.params.keys().length).toBe(0);
+		req.flush({response: [parameter]});
+		await expectAsync(responseParams).toBeResolvedTo([parameter]);
+	});
+
+	it("sends requests for a single Parameter by ID", async () => {
+		const responseParams = service.getParameters(parameter.id);
+		const req = httpTestingController.expectOne(r => r.url === `/api/${service.apiVersion}/parameters`);
+		expect(req.request.method).toBe("GET");
+		expect(req.request.params.keys().length).toBe(1);
+		expect(req.request.params.get("id")).toBe(String(parameter.id));
+		req.flush({response: [parameter]});
+		await expectAsync(responseParams).toBeResolvedTo(parameter);
+	});
+
+	it("sends requests for multiple parameters by ID", async () => {
+		const responseParams = service.getParameters(parameter.id);
+		const req = httpTestingController.expectOne(r => r.url === `/api/${service.apiVersion}/parameters`);
+		expect(req.request.method).toBe("GET");
+		expect(req.request.params.keys().length).toBe(1);
+		expect(req.request.params.get("id")).toBe(String(parameter.id));
+		const data = {
+			response: [
+				{ configFile: "test", id: 1, lastUpdated: new Date(), name: "test", secure: false, value: "test" },
+				{ configFile: "quest", id: 1, lastUpdated: new Date(), name: "quest", secure: false, value: "quest" },
+			]
+		};
+		req.flush(data);
+		await expectAsync(responseParams).toBeRejectedWithError("Traffic Ops responded with 2 Parameters by identifier 10");
+	});
+
+	it("creates new Parameters", async () => {
+		const responseParams = service.createParameter(parameter);
+		const req = httpTestingController.expectOne(`/api/${service.apiVersion}/parameters`);
+		expect(req.request.method).toBe("POST");
+		expect(req.request.params.keys().length).toBe(0);
+		expect(req.request.body).toBe(parameter);
+		req.flush({response: parameter});
+		await expectAsync(responseParams).toBeResolvedTo(parameter);
+	});
+
+	it("gets profiles associated with an existing Parameter", async () => {
+		const responseProfiles = service.getProfilesByParam(parameter);
+		const req = httpTestingController.expectOne(`/api/${service.apiVersion}/profiles?param=${parameter.id}`);
+		expect(req.request.method).toBe("GET");
+		expect(req.request.params.keys().length).toBe(1);
+		expect(req.request.body).toBe(null);
+		req.flush({response: []});
+		await expectAsync(responseProfiles).toBeResolvedTo(Array<ResponseProfile>());
+	});
+
+	it("gets profiles associated with an existing Parameter ID", async () => {
+		const responseProfiles = service.getProfilesByParam(parameter.id);
+		const req = httpTestingController.expectOne(`/api/${service.apiVersion}/profiles?param=${parameter.id}`);
+		expect(req.request.method).toBe("GET");
+		expect(req.request.params.keys().length).toBe(1);
+		expect(req.request.body).toBe(null);
+		req.flush({response: []});
+		await expectAsync(responseProfiles).toBeResolvedTo(Array<ResponseProfile>());
+	});
+
+	it("deletes existing Parameters", async () => {
+		service.deleteParameter(parameter);
+		const req = httpTestingController.expectOne(`/api/${service.apiVersion}/parameters/${parameter.id}`);
+		expect(req.request.method).toBe("DELETE");
+		expect(req.request.params.keys().length).toBe(0);
+		expect(req.request.body).toBeNull();
+		req.flush({response: parameter});
+	});
+
+	it("deletes an existing Parameter by ID", async () => {
+		service.deleteParameter(parameter.id);
+		const req = httpTestingController.expectOne(`/api/${service.apiVersion}/parameters/${parameter.id}`);
+		expect(req.request.method).toBe("DELETE");
+		expect(req.request.params.keys().length).toBe(0);
+		expect(req.request.body).toBeNull();
+		req.flush({response: parameter});
+	});
+
+	it("updates an existing Parameter", async () => {
+		const p = parameter;
+		p.value = "newValue";
+		const responseParams = service.updateParameter(parameter);
+		const req = httpTestingController.expectOne(`/api/${service.apiVersion}/parameters/${parameter.id}`);
+		expect(req.request.method).toBe("PUT");
+		expect(req.request.params.keys().length).toBe(0);
+		expect(req.request.body).toBe(p);
+		req.flush({response: p});
+		await expectAsync(responseParams).toBeResolvedTo(p);
 	});
 
 	afterEach(() => {