+++ experimental/traffic-portal/src/app/shared/navigation/navigation.service.ts
@@ -147,14 +147,25 @@ export class NavigationService {
 			}],
 			name: "Servers"
 		}, {
-			children: [{
-				href: "/core/types",
-				name: "Types"
-			},
-			{
-				href: "/core/profiles",
-				name: "Profiles"
-			}],
+			children: [
+				{
+					href: "/core/types",
+					name: "Types"
+				},
+				{
+					children: [
+						{
+							href: "/core/parameters",
+							name: "Parameters"
+						},
+						{
+							href: "/core/profiles",
+							name: "Profiles"
+						}
+					],
+					name: "Profiles"
+				}
+			],
 			name: "Configuration"
 		}, {
 			children: [{