+++ infrastructure/cdn-in-a-box/traffic_ops/config.sh
@@ -74,11 +74,6 @@ fi;
 cdn_conf=/opt/traffic_ops/app/conf/cdn.conf
 >"$cdn_conf" echo "$(jq -s '.[0] * .[1]' "$cdn_conf" <(cat <<-EOF
 {
-    "hypnotoad" : {
-        "listen" : [
-            "https://[::]?cert=$crt&key=$key&verify=0x00&ciphers=AES128-GCM-SHA256:HIGH:!RC4:!MD5:!aNULL:!EDH:!ED"
-        ]
-    },
     "disable_auto_cert_deletion": false,
     "use_ims": true,
     "server_update_status_cache_refresh_interval_sec": 0,
@@ -98,6 +93,8 @@ cdn_conf=/opt/traffic_ops/app/conf/cdn.conf
             "query_timeout_seconds": ${DEBUGGING_TIMEOUT:-60},
             "aes_key_location": "$TV_AES_KEY_LOCATION"
         },
+        "cert" : "$crt",
+        "key" : "$key",
         "proxy_timeout" : ${DEBUGGING_TIMEOUT:-60},
         "proxy_tls_timeout" : ${DEBUGGING_TIMEOUT:-60},
         "proxy_read_header_timeout" : ${DEBUGGING_TIMEOUT:-60},