+++ traffic_ops/testing/api_contract/v4/test_regions.py
@@ -29,7 +29,7 @@ Primitive = Union[bool, int, float, str, None]
 
 
 def test_region_contract(to_session: TOSession,
-	response_template_data: dict[str, Union[Primitive, list[Union[Primitive, 
+	response_template_data: dict[str, Union[Primitive, list[Union[Primitive,
 	dict[str, object], list[object]]], dict[object, object]]],
 	region_post_data: dict[str, object]
 	) -> None:
@@ -57,7 +57,7 @@ def test_region_contract(to_session: TOSession,
 
 		first_region = region_data[0]
 		if not isinstance(first_region, dict):
-			raise TypeError("malformed API response; first region in response is not an object")
+			raise TypeError("malformed API response; first region in response is not an dict")
 
 		logger.info("Region Api response %s", first_region)
 		region_response_template = response_template_data.get("regions")
@@ -78,9 +78,12 @@ def test_region_contract(to_session: TOSession,
 		pytest.fail("API contract test failed for regions endpoint: API response was malformed")
 	finally:
 		# Delete region after test execution to avoid redundancy.
-		try:
-			region_name = region_post_data["name"]
-			to_session.delete_region(query_params={"name": region_name})
-		except IndexError:
+		region_name = region_post_data.get("name")
+		if to_session.delete_region(query_params={"name": region_name}) is None:
 			logger.error("region returned by Traffic Ops is missing a 'name' property")
+			pytest.fail("Response from delete request is empty, Failing test_region_contract")
+
+		division_id = region_post_data.get("division")
+		if to_session.delete_division(division_id=division_id) is None:
+			logger.error("division returned by Traffic Ops is missing a 'name' property")
 			pytest.fail("Response from delete request is empty, Failing test_region_contract")