+++ traffic_ops/testing/api_contract/v4/test_servers.py
@@ -41,6 +41,8 @@ def test_server_contract(to_session: TOSession,
 	"""
 	# validate server keys from server get response
 	logger.info("Accessing /servers endpoint through Traffic ops session.")
+	profile_id = server_post_data[1]
+	server_post_data = server_post_data[0]
 
 	server_id = server_post_data.get("id")
 	if not isinstance(server_id, int):
@@ -57,7 +59,7 @@ def test_server_contract(to_session: TOSession,
 
 		first_server = server_data[0]
 		if not isinstance(first_server, dict):
-			raise TypeError("malformed API response; first Server in response is not an object")
+			raise TypeError("malformed API response; first Server in response is not an dict")
 		logger.info("Server Api get response %s", first_server)
 		server_response_template = response_template_data.get("servers")
 		if not isinstance(server_response_template, dict):
@@ -76,9 +78,26 @@ def test_server_contract(to_session: TOSession,
 		pytest.fail("API contract test failed for server endpoint: API response was malformed")
 	finally:
 		# Delete Server after test execution to avoid redundancy.
-		try:
-			server_id = server_post_data["id"]
-			to_session.delete_server_by_id(server_id=server_id)
-		except IndexError:
+		server_id = server_post_data.get("id")
+		if to_session.delete_server_by_id(server_id=server_id) is None:
+			logger.error("Server returned by Traffic Ops is missing an 'id' property")
+			pytest.fail("Response from delete request is empty, Failing test_server_contract")
+
+		cachegroup_id = server_post_data.get("cachegroupId")
+		if to_session.delete_cachegroups(cache_group_id=cachegroup_id) is None:
+			logger.error("cachegroup returned by Traffic Ops is missing an 'id' property")
+			pytest.fail("Response from delete request is empty, Failing test_server_contract")
+
+		if to_session.delete_profile_by_id(profile_id=profile_id) is None:
 			logger.error("Profile returned by Traffic Ops is missing an 'id' property")
+			pytest.fail("Response from delete request is empty, Failing test_server_contract")
+
+		cdn_id = server_post_data.get("cdnId")
+		if to_session.delete_cdn_by_id(cdn_id=cdn_id) is None:
+			logger.error("Cdn returned by Traffic Ops is missing an 'id' property")
+			pytest.fail("Response from delete request is empty, Failing test_server_contract")
+
+		phys_loc_id = server_post_data.get("physLocationId")
+		if to_session.delete_physical_location(physical_location_id=phys_loc_id) is None:
+			logger.error("Physical location returned by Traffic Ops is missing an 'id' property")
 			pytest.fail("Response from delete request is empty, Failing test_server_contract")